pyproject.toml
[tool.poetry]
name = "benchmark-duckdb-row-creation"
version = "0.1.0"
description = ""
authors = ["Robin Linacre <robin.linacre@digital.justice.gov.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest-benchmark = "^5.1.0"
duckdb = "^1.1.3"
pyarrow = "^18.1.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
